AVRO-861. C: Remove dependency on stdbool.h
Buildbot raised on error on Solaris, saying that the stdbool.h header is
only allowed in C99 code.  Bruce pointed out that it would be safer and
more portable to not rely on any C99 standards if we don't have to.  In
this case, it's easy enough to replace bool with int, true with 1, and
false with 0.


git-svn-id: https://svn.apache.org/repos/asf/avro/trunk@1146935 13f79535-47bb-0310-9956-ffa450edef68