AVRO-920. C: Reusable memory I/O objects
The avro_reader_t and avro_writer_t subclasses that work with memory
regions are now reusable; you can reset an existing reader or writer
object to point at a new memory region, or to simply reset to the
beginning of the current memory region.  This lets you more easily reuse
these objects when encoding and decoding large numbers of values.


git-svn-id: https://svn.apache.org/repos/asf/avro/trunk@1183635 13f79535-47bb-0310-9956-ffa450edef68