AVRO-984. C: Resolved reader initializes complex array values correctly.
Before, we weren't initialize new values when they were appended to an
array in the resolved reader class.  For arrays of primitives, this
wasn't a problem, since you don't really need to initialize primitives.
For arrays of arrays (and other complex element types), however, the
elements need to be properly initialized before they can be used.
(Contributed by Vivek Nadkarni.)


git-svn-id: https://svn.apache.org/repos/asf/avro/trunk@1224924 13f79535-47bb-0310-9956-ffa450edef68