AVRO-1088. C: Performance tests for arrays and schema resolution.
The current performance test in Avro-C measures the performance while
reading and writing of Avro values using a complex record schema, which
does not contain any arrays.

We add tests to measure the performance for simple and nested arrays. We
also replicate all tests to measure the performance of the schema
resolution using a resolved reader and a resolved writer.

Contributed by Vivek Nadkarni.


git-svn-id: https://svn.apache.org/repos/asf/avro/trunk@1340730 13f79535-47bb-0310-9956-ffa450edef68